+++ package.json
@@ -28,7 +28,6 @@
   },
   "dependencies": {
     "async": "^2.6.0",
-    "cookie": "^0.3.1",
     "most": "^1.7.2",
     "ws": "^3.3.3"
   },
