+++ packages/le-agent-session-provider/package.json
@@ -1,0 +1,36 @@
+{
+  "name": "@lp-libs/le-agent-session-provider",
+  "version": "0.1.0",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "repository": "https://github.com/busybeaver/lp-clients.git",
+  "author": {
+    "name": "busybeaver",
+    "email": "858259+busybeaver@users.noreply.github.com"
+  },
+  "contributors": [
+    {
+      "name": "busybeaver",
+      "email": "858259+busybeaver@users.noreply.github.com"
+    }
+  ],
+  "license": "Apache-2.0",
+  "scripts": {
+    "prepare": "tsc -p tsconfig.build.json"
+  },
+  "dependencies": {
+    "@lp-libs/le-session-provider": "^0.1.0",
+    "@lp-libs/util-request": "^0.1.0",
+    "@lp-libs/util-types": "^0.1.0",
+    "cookie": "^0.3.1",
+    "got": "^8.0.1"
+  },
+  "devDependencies": {
+    "typescript": "^2.7.2"
+  },
+  "engines": {
+    "node": ">=8.9.0",
+    "npm": ">=5.5.0",
+    "yarn": ">=1.5.0"
+  }
+}
