+++ packages/le-client/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./config";
+export * from "./client";
