+++ packages/le-session-provider/package.json
@@ -1,0 +1,34 @@
+{
+  "name": "@lp-libs/le-session-provider",
+  "version": "0.1.0",
+  "main": "lib/index.js",
+  "typings": "lib/index.d.ts",
+  "repository": "https://github.com/busybeaver/lp-clients.git",
+  "author": {
+    "name": "busybeaver",
+    "email": "858259+busybeaver@users.noreply.github.com"
+  },
+  "contributors": [
+    {
+      "name": "busybeaver",
+      "email": "858259+busybeaver@users.noreply.github.com"
+    }
+  ],
+  "license": "Apache-2.0",
+  "scripts": {
+    "prepare": "tsc -p tsconfig.build.json"
+  },
+  "dependencies": {
+    "@lp-libs/csds-config": "^0.1.0",
+    "@lp-libs/csds-resolver": "^0.1.0",
+    "@lp-libs/util-types": "^0.1.0"
+  },
+  "devDependencies": {
+    "typescript": "^2.7.2"
+  },
+  "engines": {
+    "node": ">=8.9.0",
+    "npm": ">=5.5.0",
+    "yarn": ">=1.5.0"
+  }
+}
